requirements.txt
Click==7.0
cycler==0.10.0
dash==1.7.0
dash-bootstrap-components==0.7.2
dash-core-components==1.6.0
dash-html-components==1.0.2
dash-renderer==1.2.2
dash-table==4.5.1
Flask==1.1.1
Flask-Compress==1.4.0
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.10.3
kafka==1.3.5
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
nltk==3.4.5
numpy==1.17.4
pandas==0.25.3
Pillow==6.2.1
pkg-resources==0.0.0
plotly==4.3.0
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.2
retrying==1.3.3
six==1.13.0
textblob==0.15.3
Werkzeug==0.16.0
wordcloud==1.6.0
